refine validate
