+++ lib/validate/schema/api-event.js
@@ -1,0 +1,42 @@
+'use strict';
+
+const apiEventSchema = {
+  '$id': '/Resources/Service/Function/Events/API',
+  'type': 'object',
+  'properties': {
+    'Type': {
+      'type': 'string',
+      'const': 'API'
+    },
+    'Properties': {
+      'type': 'object',
+      'properties': {
+        'Path': {
+          'type': 'string'
+        },
+        'Method': {
+          'type': 'string',
+          'enum': ['get', 'head', 'post', 'put', 'delete', 'connect', 'options', 'trace', 'patch',
+            'GET', 'HEAD', 'POST', 'PUT', 'DELETE', 'CONNECT', 'OPTIONS', 'TRACE', 'PATCH',
+            'Get', 'Head', 'Post', 'Put', 'Delete', 'Connect', 'Options', 'Trace', 'Patch'
+          ]
+        },
+        'RestApiId': {
+          'oneOf': [
+            { 'type': 'string' },
+            {
+              'type': 'object',
+              'properties': {
+                'Ref': { 'type': 'string' }
+              }
+            }
+          ]
+
+        }
+      },
+      'required': ['Path', 'Method']
+    }
+  }
+};
+
+module.exports = apiEventSchema;
