+++ lib/validate/schema/datahub-event.js
@@ -1,0 +1,31 @@
+'use strict';
+
+const datahubEventSchema = {
+  '$id': '/Resources/Service/Function/Events/Datahub',
+  'type': 'object',
+  'properties': {
+    'Type': {
+      'type': 'string',
+      'const': 'Datahub'
+    },
+    'Properties': {
+      'type': 'object',
+      'properties': {
+        'Topic': {
+          'type': 'string'
+        },
+        'StartingPosition': {
+          'type': 'string',
+          'enum': ['LATEST', 'OLDEST', 'SYSTEM_TIME']
+        },
+        'BatchSize': {
+          'type': 'integer',
+          'minimum': 1
+        }
+      },
+      'required': ['Topic', 'StartingPosition']
+    }
+  }
+};
+
+module.exports = datahubEventSchema;
